Complete Backend stuff
Added in information from a .json file that will be used to inject data into each cell in our grid. We are just waiting for calculated_rssi scores and actaul_rssi scores to be calculated.

The grid in html has to origin (0,0) at the top left, but in python, the origin was in the bottom right, so I had to flip the row values for all cells to properly format them into the origin being at the bottom left. The cells have all the information they needs from the .json file and are just waiting for the rssi values to be actaully in there instead of dummy data.

NOTE:
To run this code, you need a server of some sort. What I did was download an extension in Visual Studios Code called "Live Server". This can be found by clicking on the 4 cubes on the left side toolbar and entering in the words "live server". The icon should be like a purple wifi signal looking thing like this -> ((.))
Once you have it installed, you have the index.html file open and click on the bottom right of the Visual Studios Code window where it says "Go Live". After this, you should be able to view the code properly.